requirements.txt
fastapi==0.115.12
mangum==0.17.0
uvicorn==0.22.0
exceptiongroup==1.2.2
email_validator==2.2.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.1
SQLAlchemy==2.0.43
typing-inspection==0.4.1
typing_extensions==4.14.0
psycopg2==2.9.10
sqlmodel==0.0.24